Moved Edge and Node to subpackage model.fmi
